refactor(engine): extract the resume & signal-wait lifecycle to engine/resume.rs (audit 🔴-1, PR3)
Third decomposition slice of the `engine.rs` god-module. The contiguous
resume / signal-wait / cancel block moves verbatim into a sibling
`engine/resume.rs`:
- `resume_execution` — rebuilds an incomplete execution after a restart;
- `satisfy_signal_waits` / `satisfy_running_signal_waits` /
  `satisfy_signal_waits_under_lease` — deliver external signals to parked nodes;
- `cancel_dangling_nodes` / `cancel_dangling_nodes_under_lease` — tear down
  ready-but-unstarted nodes on cancellation.

They stay `impl WorkflowEngine` methods in a child module, keeping access to the
engine's private fields, sibling methods, helper free functions and types via
`use super::*` — no public API change, and no visibility bump was needed (the
moved methods are already `pub` / `pub(crate)`, and the two private ones are
only called from within the cluster).

`engine.rs`: 5,706 → 4,508 lines. Pure cut + the `mod resume;` declaration.

Verified: `cargo fmt`/clippy `-p nebula-engine --all-features --all-targets`
clean; `cargo nextest run -p nebula-engine --all-features` → 442 passed,
0 skipped (identical set to pre-move).
